+++ 1_todo/vite.config.cjs
@@ -1,0 +1,11 @@
+const { defineConfig } = require('vite');
+const { vitePlugin } = require('../src/plugins.js');
+
+module.exports = defineConfig({
+  plugins: [
+    vitePlugin({ port: 9002 })
+  ],
+  server: {
+    port: 5173
+  }
+});